+++ sw.js
@@ -1,6 +1,6 @@
 // sw.js - Thermodynamic Cache
 
-const CACHE_NAME = 'level0-v011';
+const CACHE_NAME = 'level0-v012';
 const ASSETS = [
     './',
     './index.html',